accel/amdxdna: Add hardware context management infrastructure
Add comprehensive hardware context (hwctx) management helpers for common
infrastructure shared across different hardware implementations (AIE2, AIE4).

**Resource allocation helpers (amdxdna_solver.c)**:
- amdxdna_alloc_resource(): Allocates AIE resources based on hwctx QoS
  requirements and column configuration
- amdxdna_release_resource(): Releases previously allocated AIE resources

**Column list management (amdxdna_ctx.c)**:
- amdxdna_hwctx_col_list(): Calculates and allocates column list for hwctx
  based on number of tiles and alignment requirements

**Private data lifecycle (amdxdna_ctx.c)**:
- amdxdna_hwctx_priv_init(): Initializes hwctx private data including DRM
  GPU scheduler setup, scheduler entity creation, command buffer allocation,
  and semaphore/mutex initialization
- amdxdna_hwctx_priv_fini(): Cleanup for hwctx private data - DRM scheduler
  cleanup, command buffer release, mutex destruction, column list freeing

**Complete hwctx finalization (amdxdna_ctx.c)**:
- amdxdna_hwctx_fini(): Complete hwctx cleanup that stops DRM scheduler,
  calls release_resource callback, restarts scheduler to drain pending jobs,
  waits for all jobs to complete, and destroys sync objects

**Sync object management (amdxdna_ctx.c)**:
- amdxdna_ctx_syncobj_create(): Creates DRM sync object for hwctx
- amdxdna_ctx_syncobj_destroy(): Destroys DRM sync object

**Struct reorganization**:
- Moved struct amdxdna_hwctx_priv from aie2_pci.h to amdxdna_ctx.h to enable
  common code access to private hwctx data
- Added constants: HWCTX_MAX_TIMEOUT, MAX_CHAIN_CMDBUF_SIZE

These functions provide common infrastructure for hwctx management, reducing
code duplication across different hardware implementations and improving
modularity.

Note: ARM64 builds may fail if the kernel doesn't export DRM scheduler
symbols. This is a kernel configuration issue - the code is correct for
kernels that properly export these symbols (like x86_64 upstream kernels).
