+++ drivers/accel/amdxdna/amdxdna_ctx.c
@@ -3,6 +3,7 @@
  * Copyright (C) 2022-2026, Advanced Micro Devices, Inc.
  */
 
+#include "config_kernel.h"
 #include "drm/amdxdna_accel.h"
 #include <drm/drm_device.h>
 #include <drm/drm_drv.h>
@@ -10,6 +11,7 @@
 #include <drm/drm_gem.h>
 #include <drm/drm_gem_shmem_helper.h>
 #include <drm/drm_print.h>
+#include <drm/drm_syncobj.h>
 #include <drm/gpu_scheduler.h>
 #include <linux/xarray.h>
 #include "trace/events/amdxdna.h"
@@ -508,7 +510,6 @@ int amdxdna_cmd_submit(struct amdxdna_client *client,
 		goto unlock_srcu;
 	}
 
-
 	job->hwctx = hwctx;
 	job->mm = current->mm;
 
@@ -613,4 +614,241 @@ int amdxdna_drm_submit_cmd_ioctl(struct drm_device *dev, void *data, struct drm_
 
 	XDNA_ERR(client->xdna, "Invalid command type %d", args->type);
 	return -EINVAL;
+}
+
+int amdxdna_hwctx_col_list(struct amdxdna_hwctx *hwctx, u32 row_count,
+			   u32 total_col, bool natural_align)
+{
+	struct amdxdna_dev *xdna = hwctx->client->xdna;
+	int start, end, first, last;
+	u32 width = 1, entries = 0;
+	int i;
+
+	if (!hwctx->num_tiles) {
+		XDNA_ERR(xdna, "Number of tiles is zero");
+		return -EINVAL;
+	}
+
+	if (unlikely(!row_count)) {
+		XDNA_WARN(xdna, "Core tile row count is zero");
+		return -EINVAL;
+	}
+
+	hwctx->num_col = hwctx->num_tiles / row_count;
+	if (!hwctx->num_col || hwctx->num_col > total_col) {
+		XDNA_ERR(xdna, "Invalid num_col %d", hwctx->num_col);
+		return -EINVAL;
+	}
+
+	if (natural_align)
+		width = hwctx->num_col;
+
+	/*
+	 * In range [start, end], find out columns that is multiple of width.
+	 *	'first' is the first column,
+	 *	'last' is the last column,
+	 *	'entries' is the total number of columns.
+	 */
+	start = xdna->dev_info->first_col;
+	end = total_col - hwctx->num_col;
+	if (start > 0 && end == 0) {
+		XDNA_DBG(xdna, "Force start from col 0");
+		start = 0;
+	}
+	first = start + (width - start % width) % width;
+	last = end - end % width;
+	if (last >= first)
+		entries = (last - first) / width + 1;
+	XDNA_DBG(xdna, "start %d end %d first %d last %d",
+		 start, end, first, last);
+
+	if (unlikely(!entries)) {
+		XDNA_ERR(xdna, "Start %d end %d width %d",
+			 start, end, width);
+		return -EINVAL;
+	}
+
+	hwctx->col_list = kmalloc_array(entries, sizeof(*hwctx->col_list), GFP_KERNEL);
+	if (!hwctx->col_list)
+		return -ENOMEM;
+
+	hwctx->col_list_len = entries;
+	hwctx->col_list[0] = first;
+	for (i = 1; i < entries; i++)
+		hwctx->col_list[i] = hwctx->col_list[i - 1] + width;
+
+	print_hex_dump_debug("col_list: ", DUMP_PREFIX_OFFSET, 16, 4, hwctx->col_list,
+			     entries * sizeof(*hwctx->col_list), false);
+	return 0;
+}
+
+int amdxdna_hwctx_priv_init(struct amdxdna_hwctx *hwctx,
+			    const struct drm_sched_backend_ops *sched_ops,
+			    u32 timeout_ms)
+{
+	struct amdxdna_client *client = hwctx->client;
+	struct amdxdna_dev *xdna = client->xdna;
+#ifdef HAVE_6_15_drm_sched_init
+	const struct drm_sched_init_args args = {
+		.ops = sched_ops,
+		.num_rqs = DRM_SCHED_PRIORITY_COUNT,
+		.credit_limit = HWCTX_MAX_CMDS,
+		.timeout = timeout_ms ? msecs_to_jiffies(timeout_ms) : MAX_SCHEDULE_TIMEOUT,
+		.name = "amdxdna_js",
+		.dev = xdna->ddev.dev,
+	};
+#endif
+	struct amdxdna_hwctx_priv *priv;
+	struct drm_gpu_scheduler *sched;
+	int i, ret;
+
+#ifdef HAVE_7_0_kmalloc_ops
+	priv = kzalloc_obj(*hwctx->priv);
+#else
+	priv = kzalloc(sizeof(*hwctx->priv), GFP_KERNEL);
+#endif
+	if (!priv)
+		return -ENOMEM;
+	hwctx->priv = priv;
+
+	sema_init(&priv->job_sem, HWCTX_MAX_CMDS);
+
+	for (i = 0; i < ARRAY_SIZE(priv->cmd_buf); i++) {
+		struct amdxdna_gem_obj *abo;
+		struct amdxdna_drm_create_bo args = {
+			.flags = 0,
+			.type = AMDXDNA_BO_DEV,
+			.vaddr = 0,
+			.size = MAX_CHAIN_CMDBUF_SIZE,
+		};
+
+		abo = amdxdna_drm_create_dev_bo(&xdna->ddev, &args, client->filp);
+		if (IS_ERR(abo)) {
+			ret = PTR_ERR(abo);
+			goto free_cmd_bufs;
+		}
+
+		XDNA_DBG(xdna, "Command buf %d addr 0x%llx size 0x%lx",
+			 i, abo->mem.dma_addr, abo->mem.size);
+		priv->cmd_buf[i] = abo;
+	}
+
+	sched = &priv->sched;
+	mutex_init(&priv->io_lock);
+
+	fs_reclaim_acquire(GFP_KERNEL);
+	might_lock(&priv->io_lock);
+	fs_reclaim_release(GFP_KERNEL);
+#ifdef HAVE_6_15_drm_sched_init
+	ret = drm_sched_init(sched, &args);
+#else
+	ret = drm_sched_init(sched, sched_ops, NULL, DRM_SCHED_PRIORITY_COUNT,
+			     HWCTX_MAX_CMDS, 0,
+			     timeout_ms ? timeout_ms : MAX_SCHEDULE_TIMEOUT,
+			     NULL, NULL, "amdxdna_js", xdna->ddev.dev);
+#endif
+	if (ret) {
+		XDNA_ERR(xdna, "Failed to init DRM scheduler. ret %d", ret);
+		goto free_cmd_bufs;
+	}
+
+	ret = drm_sched_entity_init(&priv->entity, DRM_SCHED_PRIORITY_NORMAL,
+				    &sched, 1, NULL);
+	if (ret) {
+		XDNA_ERR(xdna, "Failed to initial sched entiry. ret %d", ret);
+		goto free_sched;
+	}
+
+	return 0;
+
+free_sched:
+	drm_sched_fini(&priv->sched);
+free_cmd_bufs:
+	for (i = 0; i < ARRAY_SIZE(priv->cmd_buf); i++) {
+		if (!priv->cmd_buf[i])
+			break;
+		drm_gem_object_put(to_gobj(priv->cmd_buf[i]));
+	}
+	kfree(priv);
+	return ret;
+}
+
+void amdxdna_hwctx_priv_fini(struct amdxdna_hwctx *hwctx)
+{
+	struct amdxdna_hwctx_priv *priv = hwctx->priv;
+	int idx;
+
+	drm_sched_fini(&priv->sched);
+
+	for (idx = 0; idx < ARRAY_SIZE(priv->cmd_buf); idx++)
+		drm_gem_object_put(to_gobj(priv->cmd_buf[idx]));
+
+	mutex_destroy(&priv->io_lock);
+	kfree(hwctx->col_list);
+	kfree(priv);
+}
+
+void amdxdna_hwctx_fini(struct amdxdna_hwctx *hwctx,
+			void (*release_resource)(struct amdxdna_hwctx *hwctx))
+{
+	struct amdxdna_dev *xdna = hwctx->client->xdna;
+
+	/* Stop scheduler, release hw resource, restart to drain pending jobs */
+	drm_sched_stop(&hwctx->priv->sched, NULL);
+	release_resource(hwctx);
+#ifdef HAVE_6_13_drm_sched_start_errno
+	drm_sched_start(&hwctx->priv->sched, 0);
+#else
+	drm_sched_start(&hwctx->priv->sched);
+#endif
+
+	mutex_unlock(&xdna->dev_lock);
+	drm_sched_entity_destroy(&hwctx->priv->entity);
+
+	/* Wait for all submitted jobs to be completed or canceled */
+	wait_event(hwctx->priv->job_free_wq,
+		   atomic64_read(&hwctx->job_submit_cnt) ==
+		   atomic64_read(&hwctx->job_free_cnt));
+	mutex_lock(&xdna->dev_lock);
+
+	amdxdna_ctx_syncobj_destroy(hwctx);
+	amdxdna_hwctx_priv_fini(hwctx);
+}
+
+int amdxdna_ctx_syncobj_create(struct amdxdna_hwctx *hwctx)
+{
+	struct amdxdna_dev *xdna = hwctx->client->xdna;
+	struct drm_syncobj *syncobj;
+	struct drm_file *filp;
+	u32 hdl;
+	int ret;
+
+	filp = hwctx->client->filp;
+
+	hwctx->syncobj_hdl = AMDXDNA_INVALID_FENCE_HANDLE;
+
+	ret = drm_syncobj_create(&syncobj, 0, NULL);
+	if (ret) {
+		XDNA_ERR(xdna, "Create ctx syncobj failed, ret %d", ret);
+		return ret;
+	}
+	ret = drm_syncobj_get_handle(filp, syncobj, &hdl);
+	if (ret) {
+		drm_syncobj_put(syncobj);
+		XDNA_ERR(xdna, "Create ctx syncobj handle failed, ret %d", ret);
+		return ret;
+	}
+	hwctx->priv->syncobj = syncobj;
+	hwctx->syncobj_hdl = hdl;
+
+	return 0;
+}
+
+void amdxdna_ctx_syncobj_destroy(struct amdxdna_hwctx *hwctx)
+{
+	/*
+	 * The syncobj_hdl is owned by user space and will be cleaned up
+	 * separately.
+	 */
+	drm_syncobj_put(hwctx->priv->syncobj);
 }
